pyproject.toml
[project]
name = "catroslot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.2.1",
]

[dependency-groups]
dev = [
    "sphinx>=8.2.3",
]
